More sonarcloud and fixes
